structural code refactor
