+++ src/frame/utils/getAdjustedPositionSize.js
