+++ src/frame/utils/index.js
@@ -1,8 +1,8 @@
 import getAdjustedPositionSize from './getAdjustedPositionSize';
 import toMarginGraphic from './toMarginGraphic';
 import generateFrameTitle from './generateFrameTitle';
-import getExtent from '../plots/BasePlot/getExtent';
-import toPipeline from './toPipeline';
+import getExtent from '../../plots/BasePlot/getExtent';
+import toPipeline from '../toPipeline';
 
 export {
   getAdjustedPositionSize,
